% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chi_square.test}
\alias{chi_square.test}
\title{Perform a Chi-square test for a given sample variance}
\usage{
chi_square.test(
  var,
  sigma2,
  n,
  alternative = c("two-sided", "greater", "less"),
  confint = F,
  alpha = 0.05
)
}
\arguments{
\item{var}{The population variance}

\item{sigma2}{The sample variance}

\item{n}{Sample size (default is 1).}

\item{alternative}{Type of alternative hypothesis ("two-sided," "less," or "greater").}

\item{confint}{Whether to calculate a confidence interval (default is FALSE).}

\item{alpha}{Significance level (default is 0.05).}
}
\value{
The test result (reject or fail to reject null hypothesis) and an optional confidence interval.
}
\description{
This function conducts a Chi-square test based on a sample variance You can specify the
population variance, sample size, alternative hypothesis, significance level, and whether to
calculate a confidence interval.
}
