% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paired_diff.test}
\alias{paired_diff.test}
\title{Perform a Z-test for a given sample mean}
\usage{
paired_diff.test(
  diff_vec = F,
  vec1 = F,
  vec2 = F,
  alternative = c("two-sided", "greater", "less"),
  confint = F,
  alpha = 0.05
)
}
\arguments{
\item{diff_vec}{The difference vector.}

\item{vec1}{First set of data.}

\item{vec2}{Second set of data.}

\item{alternative}{Type of alternative hypothesis ("two-sided," "less," or "greater").}

\item{confint}{Whether to calculate a confidence interval (default is FALSE).}

\item{alpha}{Significance level (default is 0.05).}
}
\value{
The test result (reject or fail to reject null hypothesis) and an optional confidence interval.
}
\description{
This function conducts a Z-test based on a sample mean. You can specify the population mean,
standard deviation, sample size, alternative hypothesis, significance level, and whether to
calculate a confidence interval.
}
