% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop.test}
\alias{prop.test}
\title{Perform a Z-test for a given proportion}
\usage{
prop.test(
  x,
  n,
  pi,
  alternative = c("two-sided", "greater", "less"),
  confint = F,
  alpha = 0.05
)
}
\arguments{
\item{x}{The number of favorable observations.}

\item{n}{Sample size.}

\item{pi}{population proportion to be tested}

\item{alternative}{Type of alternative hypothesis ("two-sided," "less," or "greater").}

\item{confint}{Whether to calculate a confidence interval (default is FALSE).}

\item{alpha}{Significance level (default is 0.05).}
}
\value{
The test result (reject or fail to reject null hypothesis) and an optional confidence interval.
}
\description{
This function conducts a Z-test based on a sample prop. You can specify the population mean,
standard deviation, sample size, alternative hypothesis, significance level, and whether to
calculate a confidence interval.
}
