% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t.test}
\alias{t.test}
\title{Perform a T-test for a given sample mean}
\usage{
\method{t}{test}(
  x,
  mu = 0,
  sd = 1,
  n = 2,
  alternative = c("two-sided", "less", "greater"),
  alpha = 0.05,
  confint = F
)
}
\arguments{
\item{x}{The sample mean.}

\item{mu}{The population mean (default is 0).}

\item{sd}{The standard deviation of the population (default is 1).}

\item{n}{Sample size (default is 1).}

\item{alternative}{Type of alternative hypothesis ("two-sided," "less," or "greater").}

\item{alpha}{Significance level (default is 0.05).}

\item{confint}{Whether to calculate a confidence interval (default is FALSE).}
}
\value{
The test result (reject or fail to reject null hypothesis) and an optional confidence interval.
}
\description{
This function conducts a T-test based on a sample mean. You can specify the population mean,
standard deviation, sample size, alternative hypothesis, significance level, and whether to
calculate a confidence interval.
}
