% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{two_sample.z_test}
\alias{two_sample.z_test}
\title{Perform a two-sample Z-test for given sample means}
\usage{
two_sample.z_test(
  x1,
  x2,
  s1,
  s2,
  n1,
  n2,
  alternative = c("two-sided", "greater", "less"),
  confint = F,
  alpha = 0.05,
  equal_var = F
)
}
\arguments{
\item{x1}{The mean of sample 1.}

\item{x2}{The mean of sample 2.}

\item{s1}{The standard deviation of the sample 1.}

\item{s2}{The standard deviation of the sample 2.}

\item{n1}{Size of sample 1.}

\item{n2}{Size of sample 2.}

\item{alternative}{Type of alternative hypothesis ("two-sided," "less," or "greater").}

\item{confint}{Whether to calculate a confidence interval (default is FALSE).}

\item{alpha}{Significance level (default is 0.05).}
}
\value{
The test result (reject or fail to reject null hypothesis) and an optional confidence interval.
}
\description{
This function conducts a Z-test based on sample means. You can specify the sample means,
standard deviations, sample sizes, alternative hypothesis, significance level, and whether to
calculate a confidence interval.
}
